Am rganizat functiile pe clase si pachete astfel incat sa implementeze MVC
