+++ src/main/java/controller/servlets/ServletTest.java
@@ -1,4 +1,4 @@
-package servlets;
+package controller.servlets;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
