+++ src/main/test/ProblemTest.java
@@ -1,21 +1,19 @@
-import controller.Problem;
+import Model.dao.Problem;
 import org.junit.Test;
 
-import java.sql.SQLException;
-
-import static org.junit.Assert.*;
-
+import static org.junit.Assert.assertEquals;
+import static Model.dao.Test.*;
+import static Model.dao.User.*;
 public class ProblemTest {
 
     @Test
     public void getPunctaj() {
-        Problem pr = new Problem();
-        assertEquals(3, pr.getUserProblemScore(1, 5)); //trebuie adaugate punctaje in tabela pentru a putea testa
+        assertEquals(3, getUserProblemScore(1, 5)); //trebuie adaugate punctaje in tabela pentru a putea testa
     }
 
     @Test
     public void getPunctajNull() {
         Problem pr = new Problem();
-        assertEquals(-1, pr.getUserProblemScore(2, 3));
+        assertEquals(-1, getUserProblemScore(2, 3));
     }
 }
